Add a videos page header and switch the videos and shows pages
